requirements.txt
python-dotenv
jupyterlab
ipywidgets
urllib3
openai
gradio
google.generativeai
anthropic
groq
black
pymongo
pandas
python-chess
